+++ src/renderer/components/savers/Saver.utils.ts
@@ -3,7 +3,7 @@ import { Asset, BaseAmount, baseAmount } from '@xchainjs/xchain-util'
 import { ZERO_BASE_AMOUNT } from '../../const'
 import { isChainAsset, max1e8BaseAmount } from '../../helpers/assetHelper'
 import { getChainAsset } from '../../helpers/chainHelper'
-import { SaverDepositFees } from '../../services/chain/types'
+import { SaverDepositFees, WithdrawAssetFees } from '../../services/chain/types'
 
 /**
  * Calculates max. balance available to swap
@@ -40,4 +40,14 @@ export const getZeroSaverDepositFees = (asset: Asset): SaverDepositFees => ({
     outFee: ZERO_BASE_AMOUNT,
     refundFee: ZERO_BASE_AMOUNT
   }
+})
+
+/**
+ * Returns zero sym deposit fees
+ * by given paired asset to deposit
+ */
+export const getZeroSaverWithdrawFees = (asset: Asset): WithdrawAssetFees => ({
+  asset,
+  inFee: ZERO_BASE_AMOUNT,
+  outFee: ZERO_BASE_AMOUNT
 })