+++ src/renderer/i18n/en/savers.ts
@@ -5,7 +5,19 @@ const savers: SaversMessages = {
   'savers.detail.title': 'Your savings',
   'savers.detail.current.title': 'Deposit value',
   'savers.detail.redeem.title': 'Redeem value',
-  'savers.detail.percent': 'Growth'
+  'savers.detail.percent': 'Growth',
+  'savers.info.max.redeem.value': 'Max value to redeem',
+  'savers.info.max.balance': 'Max balance',
+  'savers.add.state.sending': 'Sending add saver transaction',
+  'savers.add.state.checkResults': 'Checking transaction results',
+  'savers.add.state.pending': 'Adding saver',
+  'savers.add.state.success': 'Successful saver add',
+  'savers.withdraw.state.sending': 'Sending withdraw saver transaction',
+  'savers.withdraw.state.checkResults': 'Checking transaction results',
+  'savers.withdraw.state.pending': 'Withdrawing Saver',
+  'savers.withdraw.state.success': 'Succesful saver withdraw',
+  'savers.add.state.error': 'Error adding savers',
+  'savers.withdraw.state.error': 'Error withdrawing savers'
 }
 
 export default savers