+++ src/renderer/i18n/fr/savers.ts
@@ -5,7 +5,19 @@ const savers: SaversMessages = {
   'savers.detail.title': 'Vos économies',
   'savers.detail.current.title': 'Valeur du dépôt',
   'savers.detail.redeem.title': 'Valeur de rachat',
-  'savers.detail.percent': 'Croissance'
+  'savers.detail.percent': 'Croissance',
+  'savers.info.max.redeem.value': 'Valeur maximale à échanger',
+  'savers.info.max.balance': 'Solde maximal',
+  'savers.add.state.sending': 'Envoi de la transaction d’ajout de l’épargnant',
+  'savers.add.state.checkResults': 'Vérification des résultats de la transaction',
+  'savers.add.state.pending': 'Ajout de l’épargnant en cours',
+  'savers.add.state.success': 'Ajout de l’épargnant réussi',
+  'savers.withdraw.state.sending': 'Envoi de la transaction de retrait de l’épargnant',
+  'savers.withdraw.state.checkResults': 'Vérification des résultats de la transaction',
+  'savers.withdraw.state.pending': 'Retrait de l’épargnant en cours',
+  'savers.withdraw.state.success': 'Retrait de l’épargnant réussi',
+  'savers.add.state.error': 'Erreur lors de l’ajout de l’épargnant',
+  'savers.withdraw.state.error': 'Erreur lors du retrait de l’épargnant'
 }
 
 export default savers