+++ src/renderer/i18n/ru/savers.ts
@@ -5,7 +5,19 @@ const savers: SaversMessages = {
   'savers.detail.title': 'Твои сбережения',
   'savers.detail.current.title': 'Сколько внесено',
   'savers.detail.redeem.title': 'Сколько заберёшь',
-  'savers.detail.percent': 'Рост'
+  'savers.detail.percent': 'Рост',
+  'savers.info.max.redeem.value': 'Максимальная сумма для вывода',
+  'savers.info.max.balance': 'Максимальный баланс',
+  'savers.add.state.sending': 'Отправка транзакции на добавление сберегателя',
+  'savers.add.state.checkResults': 'Проверка результатов транзакции',
+  'savers.add.state.pending': 'Добавление сберегателя',
+  'savers.add.state.success': 'Успешное добавление сберегателя',
+  'savers.withdraw.state.sending': 'Отправка транзакции на снятие сберегателя',
+  'savers.withdraw.state.checkResults': 'Проверка результатов транзакции',
+  'savers.withdraw.state.pending': 'Снятие сберегателя',
+  'savers.withdraw.state.success': 'Успешное снятие сберегателя',
+  'savers.add.state.error': 'Ошибка при добавлении сберегателя',
+  'savers.withdraw.state.error': 'Ошибка при снятии сберегателя'
 }
 
 export default savers