+++ src/renderer/i18n/types.ts
@@ -482,6 +482,18 @@ type SaversMessageKey =
   | 'savers.detail.current.title'
   | 'savers.detail.redeem.title'
   | 'savers.detail.percent'
+  | 'savers.info.max.redeem.value'
+  | 'savers.info.max.balance'
+  | 'savers.add.state.sending'
+  | 'savers.add.state.checkResults'
+  | 'savers.add.state.pending'
+  | 'savers.add.state.success'
+  | 'savers.add.state.error'
+  | 'savers.withdraw.state.sending'
+  | 'savers.withdraw.state.checkResults'
+  | 'savers.withdraw.state.pending'
+  | 'savers.withdraw.state.success'
+  | 'savers.withdraw.state.error'
 
 export type SaversMessages = { [key in SaversMessageKey]: string }