+++ src/renderer/views/savers/SaversView.tsx
@@ -29,7 +29,9 @@ import { useWalletContext } from '../../contexts/WalletContext'
 import { getAssetFromNullableString } from '../../helpers/assetHelper'
 import { eqChain, eqNetwork, eqWalletType } from '../../helpers/fp/eq'
 import { sequenceTOption, sequenceTRD } from '../../helpers/fpHelpers'
+import * as PoolHelpers from '../../helpers/poolHelper'
 import { addressFromOptionalWalletAddress } from '../../helpers/walletHelper'
+import { useMimirHalt } from '../../hooks/useMimirHalt'
 import { useNetwork } from '../../hooks/useNetwork'
 import { useOpenExplorerTxUrl } from '../../hooks/useOpenExplorerTxUrl'
 import { usePricePool } from '../../hooks/usePricePool'
@@ -80,8 +82,9 @@ const Content: React.FC<Props> = (props): JSX.Element => {
   const { network } = useNetwork()
   const { reloadInboundAddresses } = useThorchainContext()
 
-  const { reloadSaverProvider } = useThorchainContext()
-  const { assetWithDecimal$, addressByChain$, reloadSaverDepositFee, saverDeposit$ } = useChainContext()
+  const { getSaverProvider$, reloadSaverProvider } = useThorchainContext()
+
+  const { assetWithDecimal$, addressByChain$, reloadSaverDepositFee, saverDeposit$, saverWithdraw$ } = useChainContext()
   const { approveERC20Token$, isApprovedERC20Token$, approveFee$, reloadApproveFee } = useEthereumContext()
   const {
     balancesState$,
@@ -94,12 +97,14 @@ const Content: React.FC<Props> = (props): JSX.Element => {
 
   const {
     service: {
-      pools: { poolsState$, reloadPools, reloadSelectedPoolDetail, selectedPoolAddress$ },
+      pools: { poolsState$, reloadPools, reloadSelectedPoolDetail, selectedPoolAddress$, haltedChains$ },
       setSelectedPoolAsset
     }
   } = useMidgardContext()
   const oPoolAddress: O.Option<PoolAddress> = useObservableState(selectedPoolAddress$, O.none)
 
+  const [haltedChains] = useObservableState(() => FP.pipe(haltedChains$, RxOp.map(RD.getOrElse((): Chain[] => []))), [])
+  const { mimirHalt } = useMimirHalt()
   // reload inbound addresses at `onMount` to get always latest `pool address` + `feeRates`
   useEffect(() => {
     reloadInboundAddresses()
@@ -255,6 +260,15 @@ const Content: React.FC<Props> = (props): JSX.Element => {
                 assetDetails,
                 A.map(({ asset }) => asset)
               )
+              const disableAllPoolActions = (chain: Chain) =>
+                PoolHelpers.disableAllActions({ chain, haltedChains, mimirHalt })
+
+              const disableTradingPoolActions = (chain: Chain) =>
+                PoolHelpers.disableTradingActions({ chain, haltedChains, mimirHalt })
+
+              const checkDisableSaverAction = () => {
+                return disableAllPoolActions(chain) || disableTradingPoolActions(chain)
+              }
               const getTabContentByIndex = (index: number) => {
                 switch (index) {
                   case TabIndex.ADD:
@@ -284,16 +298,38 @@ const Content: React.FC<Props> = (props): JSX.Element => {
                         saverDeposit$={saverDeposit$}
                         hidePrivateData={false}
                         onChangeAsset={onChangeAssetHandler}
+                        disableSaverAction={checkDisableSaverAction()}
                       />
                     )
                   case TabIndex.WITHDRAW:
                     return (
                       <WithdrawSavers
+                        keystore={keystore}
+                        poolDetails={poolDetails}
+                        asset={new CryptoAmount(baseAmount(0, assetWD.decimal), assetWD.asset)}
+                        walletBalances={balancesState}
                         network={network}
-                        asset={assetWD}
                         pricePool={pricePool}
                         fees$={saverWithdrawFee$}
                         address={address}
+                        validatePassword$={validatePassword$}
+                        goToTransaction={openExplorerTxUrl}
+                        getExplorerTxUrl={getExplorerTxUrl}
+                        sourceWalletType={walletType}
+                        poolAssets={poolAssets}
+                        reloadBalances={reloadHandler}
+                        approveFee$={approveFee$}
+                        reloadApproveFee={reloadApproveFee}
+                        reloadFees={reloadSaverDepositFee}
+                        reloadSelectedPoolDetail={reloadSelectedPoolDetail}
+                        isApprovedERC20Token$={isApprovedERC20Token$}
+                        approveERC20Token$={approveERC20Token$}
+                        poolAddress={oPoolAddress}
+                        saverWithdraw$={saverWithdraw$}
+                        hidePrivateData={false}
+                        onChangeAsset={onChangeAssetHandler}
+                        saverPosition={getSaverProvider$}
+                        disableSaverAction={checkDisableSaverAction()}
                       />
                     )
                   default: